HMW-45 Removed the EsriModules context.
